qcacld-2.0: Prevent runtime suspend when recovery initiated
When self recovery initiated, CRASH_INJECT command is sent to
firmware. If firmware crash doesn't reach the host in time then
RuntimePM may kick in and try to suspend the PCIe link. At this
time, firmware might already been crashed just that host missed
the interrupt. Add the logic to prevent the Runtime suspend when
initiating the recovery so that host driver doesn't attempt the
suspend which may not go through and cause stability issues.

CRs-fixed: 924538
Git-commit: 01ae45b557e662b39fcdf41eed2bfc804ae5bd99
Bug: 23785794
Signed-off-by: Amarnath Hullur Subramanyam <amarnath@codeaurora.org>