qcacld-2.0 Reduce LL stats / Firmware Memory Dump Blocking time
prima to qcacld-2.0 propagation.

Currently driver block wpa_supplicant context for response
for 5 / 5 seconds for LL stats and firmware memory dump
feature respectively. But in M framework watch dog will
be triggered after 1 second because driver is holding
the context longer, which lead to wd thread starvation.
As a part of this fix, block time is reduced to 800 msec
respectively in both cases.

Change-Id: I558a7e40242d04d7a909a819cbdcdd46d0e894cf
CRs-Fixed: 951119
Bug: 28605342
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>