qcacld-2.0: Prevent wlan exit in wrong state
When Wifi is turned off, then hdd_stop is called
which leads to unload of driver. But before hdd_stop
is called, user has turned on wifi and
fwpath_change_handler is called before hdd_stop which
leads to unload twice and leading to crash.

To avoid this problem,
1) Synchronize the kickstart_driver call to keep
wlan_hdd_inited updated.

2) Store the fwpath string locally and check the last mode
with the current mode in fwpath_change_handler.
- If load request comes and driver is already
  loaded with no change in fwpath string, then unload the
  driver and return failure.
- If load request comes and driver is already loaded with change
  in fwpath string then unload and load the driver again.

Change-Id: I2b22a0280e6a408def30244d5cd8ac306425b0bb
CRs-Fixed: 1012569
Bug: 28402915
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>