new: adding a h818 kernel conf
the h818 can boot custom kernels only with UsU but UsU
will bring up an incomplete bootloader cmdline which
then results in TOUCH is not working for the h818 and
ONLY for that model.
The reason is that a h818 will init the power for the
TOUCH screen only when a special parameter is found in
the boot cmdline.
The kernel config will completely OVERRIDE the cmdline
to ensure all is in place (mainly required is: lge.sim_num=2)
and the model is properly set.

Thanks for that great finding goes 100% to scholbert
(https://forum.xda-developers.com/showpost.php?p=75307864&postcount=370)

Change-Id: Iedb62e2b15a3b314b94e22fa52735e39339798af