dts: fix wrong memory reservation in pstore
in some cases a kernel panic was triggered due to an overlapping
memory reservation due to a wrong mem address.

and ... well I was maybe drunk or so but this takes back the rest
of the memory nodes which fixes all related stuff as well ofc.

this fixes also the failure on boot for another reservation request:

> ramoops: attached 0x200000@0x1fe00000, ecc: 16/0
> request mem region (0x20000@0xffc0000) failed

Change-Id: Id37327fd916c960c0b9007940d48ea0ea2a1955c