defconfig: enable auto-reboot on kernel panic
the MSM download mode (the demigod screen) prevents the auto-rebooting,
besides that it is useful only when having the proper tools / knowledge
(which, sorry, the average user does not have),
and last but not least rebooting by using volume button doesn't work in that mode (often).

this commit sets the auto-reboot to 15s which should be more then enough
to write all panic data before.